# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for authenticated Linux users to manage Containerlab deployments, topology files, node access, captures, and host networking tools.\n\nMost endpoints under /api/v1 require a JWT from POST /login. Send it as: Authorization: Bearer \u003ctoken\u003e.",
        "title": "Containerlab API Server",
        "contact": {
            "name": "Containerlab API Server maintainers",
            "url": "https://github.com/srl-labs/clab-api-server/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/labs/workspace/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Streams create/change/delete/rename events inside the authenticated user's editable lab workspace root as NDJSON.\nInternal `.archive-*` and `.backup-*` transaction paths are excluded from the stream.",
                "produces": [
                    "application/x-ndjson"
                ],
                "tags": [
                    "Labs"
                ],
                "summary": "Stream lab workspace file events",
                "responses": {
                    "200": {
                        "description": "Workspace file event stream",
                        "schema": {
                            "$ref": "#/definitions/models.WorkspaceFileEventResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid credentials or user not in allowed group"
                }
            }
        },
...
        "models.WorkspaceFileEventResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "delete"
                },
                "kind": {
                    "type": "string",
                    "example": "file"
                },
                "parentPath": {
                    "type": "string",
                    "example": "lab1/configs"
                },
                "path": {
                    "type": "string",
                    "example": "lab1/configs/startup.cfg"
                },
                "type": {
                    "type": "string",
                    "example": "workspace-file"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}